refactor: useAuth 훅 재구현
