+++ client/src/pages/Lottery/index.tsx
@@ -1,11 +1,8 @@
-import { useCallback, useEffect, useState } from "react";
-import { useCookies } from "react-cookie";
-import { useLoaderData, useLocation, useNavigate } from "react-router-dom";
-import { AuthAPI } from "@/apis/authAPI";
+import { useCallback } from "react";
+import { useLoaderData } from "react-router-dom";
 import Footer from "@/components/Footer";
 import Notice from "@/components/Notice";
 import { LOTTERY_SECTIONS } from "@/constants/PageSections/sections.ts";
-import { COOKIE_KEY } from "@/constants/cookie";
 import {
     CustomDesign,
     HeadLamp,
@@ -17,16 +14,12 @@ import {
     SmileBadge,
     WheelDesign,
 } from "@/features/Lottery";
-import useFetch from "@/hooks/useFetch";
+import { useAuth } from "@/hooks/useAuth.ts";
 import useHeaderStyleObserver from "@/hooks/useHeaderStyleObserver.ts";
-import usePhoneNumberDispatchContext from "@/hooks/usePhoneNumberDispatchContext";
-import usePhoneNumberStateContext from "@/hooks/usePhoneNumberStateContext";
 import usePopup from "@/hooks/usePopup";
 import useScrollTop from "@/hooks/useScrollTop";
 import useToast from "@/hooks/useToast";
-import { PostAuthResponse } from "@/types/authApi";
 import { GetLotteryResponse } from "@/types/lotteryApi";
-import { PHONE_NUMBER_ACTION } from "@/types/phoneNumber";
 import { getMsTime } from "@/utils/getMsTime";
 
 export default function Lottery() {
@@ -35,49 +28,10 @@ export default function Lottery() {
         darkSections: [LOTTERY_SECTIONS.HEADLINE, LOTTERY_SECTIONS.SHORT_CUT],
     });
 
-    const navigate = useNavigate();
-    const location = useLocation();
-    const queryParams = new URLSearchParams(location.search);
-    const inviteUser = queryParams.get(COOKIE_KEY.INVITE_USER);
-
     const lotteryData = useLoaderData() as GetLotteryResponse;
 
-    const [_cookies, setCookie] = useCookies([COOKIE_KEY.ACCESS_TOKEN, COOKIE_KEY.INVITE_USER]);
-
-    const {
-        data: authToken,
-        isSuccess: isSuccessGetAuthToken,
-        fetchData: getAuthToken,
-    } = useFetch<PostAuthResponse, string>((val: string) =>
-        AuthAPI.getAuthToken({ phoneNumber: val })
-    );
-
-    const { phoneNumber } = usePhoneNumberStateContext();
-    const dispatch = usePhoneNumberDispatchContext();
-
-    const [phoneNumberState, setPhoneNumberState] = useState(phoneNumber);
-
-    useEffect(() => {
-        if (inviteUser) {
-            setCookie(COOKIE_KEY.INVITE_USER, inviteUser);
-        }
-    }, [inviteUser]);
-
-    useEffect(() => {
-        if (authToken && isSuccessGetAuthToken) {
-            setCookie(COOKIE_KEY.ACCESS_TOKEN, authToken.accessToken);
-            dispatch({ type: PHONE_NUMBER_ACTION.SET_PHONE_NUMBER, payload: phoneNumberState });
-            navigate("/lottery/custom");
-        }
-    }, [authToken, isSuccessGetAuthToken]);
-
-    const handlePhoneNumberChange = (val: string) => {
-        setPhoneNumberState(val);
-    };
-
-    const handlePhoneNumberConfirm = async (val: string) => {
-        await getAuthToken(val);
-    };
+    const { phoneNumberState, handlePhoneNumberChange, handlePhoneNumberConfirm } =
+        useAuth("/lottery/custom");
 
     const { handleOpenPopup, PopupComponent } = usePopup({
         phoneNumber: phoneNumberState,
